Work on the replication test.

+++ db4oj/test/src/com/db4o/test/replication/ReplicationFeaturesMain.java
@@ -12,13 +12,6 @@ import com.db4o.test.*;
 
 public class ReplicationFeaturesMain {
 
-	
-	
-	// THIS IS WORK IN PROGRESS BY KLAUS.
-	
-	
-	
-	
 	private static final Set A = new HashSet(1); 
 	private static final Set B = new HashSet(1); 
 	private static final Set BOTH = new HashSet(2); 
@@ -56,26 +49,21 @@ public class ReplicationFeaturesMain {
     }
 	
     public void test() {
+        
+        if (1 == 1) return;
+
+        
+        System.out.println("This strategy is too complex. I will try a different one.");
+        System.out.println("WORK IN PROGRESS BY KLAUS.");
+
+        
 		_a = Test.objectContainer();
 		_b = Test.replica();
 		deleteAll(_a);
 		deleteAll(_b);
 
-		clearNameExpectations();
-		addNames(_possibleNamesInA, ALL_POSSIBLE_NAMES);
-		addNames(_possibleNamesInB, INITIAL_NAMES);
-		addNames(_impossibleNamesInB, NAME_CHANGES);
 		tstDirection(A);
-
-		clearNameExpectations();
-		addNames(_possibleNamesInA, INITIAL_NAMES);
-		addNames(_impossibleNamesInA, NAME_CHANGES);
-		addNames(_possibleNamesInB, ALL_POSSIBLE_NAMES);
 		tstDirection(B);
-		
-		clearNameExpectations();
-		addNames(_possibleNamesInA, ALL_POSSIBLE_NAMES);
-		addNames(_possibleNamesInB, ALL_POSSIBLE_NAMES);
 		tstDirection(BOTH);
 		
 //      TODO: replication.checkConflict(obj); //(peek)
@@ -125,16 +113,78 @@ public class ReplicationFeaturesMain {
 	}
 
 	private void doIt() {
-		initState();
+
+        initState();
+        initNameExpectations();
+        
+        if (_direction.contains("A")) {
+            addNames(_possibleNamesInA, NAME_CHANGES);
+        } else {
+            addNames(_impossibleNamesInA, NAME_CHANGES);
+        }
+
+        if (_direction.contains("B")) {
+            addNames(_possibleNamesInB, NAME_CHANGES);
+        } else {
+            addNames(_impossibleNamesInB, NAME_CHANGES);
+        }
+
+        
+        if (!_containersWithNewObjects.contains("A")) {
+            _impossibleNamesInA.add("newInA");
+            _impossibleNamesInB.add("newInA");
+        }
+
+        if (!_containersWithNewObjects.contains("B")) {
+            _impossibleNamesInA.add("newInB");
+            _impossibleNamesInB.add("newInB");
+        }
+
+        
+        if (_containersWithChangedObjects.contains("A")) {
+            changeObject(_a, "a1", "a1ChangedInA");
+            changeObject(_a, "b1", "b1ChangedInA");
+            _impossibleNamesInA.add("a1");
+            _impossibleNamesInA.add("b1");
+        } else {
+            _impossibleNamesInA.add("a1ChangedInA");
+            _impossibleNamesInA.add("b1ChangedInA");
+            _impossibleNamesInB.add("a1ChangedInA");
+            _impossibleNamesInB.add("b1ChangedInA");
+        }
+
+        if (_containersWithChangedObjects.contains("B")) {
+            changeObject(_b, "a1", "a1ChangedInB");
+            changeObject(_b, "b1", "b1ChangedInB");
+            _impossibleNamesInB.add("a1");
+            _impossibleNamesInB.add("b1");
+        } else {
+            _impossibleNamesInA.add("a1ChangedInB");
+            _impossibleNamesInA.add("b1ChangedInB");
+            _impossibleNamesInB.add("a1ChangedInB");
+            _impossibleNamesInB.add("b1ChangedInB");
+        }
+        
+        
+        _possibleNamesInA.removeAll(_impossibleNamesInA);
+        _possibleNamesInB.removeAll(_impossibleNamesInB);
 		checkNames(_a, _possibleNamesInA);
 		checkNames(_b, _possibleNamesInB);
-		checkClean();
-		
 	}
 
-	private void clearNameExpectations() {
+    private void changeObject(ObjectContainer container, String name, String newName) {
+        Replicated obj = find(container, name);
+        obj._name = newName;
+        container.set(obj);
+    }
+
+    private void initNameExpectations() {
 		_possibleNamesInA = new HashSet();
+        addNames(_possibleNamesInA, INITIAL_NAMES);
+        
 		_possibleNamesInB = new HashSet();
+        addNames(_possibleNamesInB, INITIAL_NAMES);
+        
 		_impossibleNamesInA = new HashSet();
 		_impossibleNamesInB = new HashSet();
 	}
@@ -146,32 +196,33 @@ public class ReplicationFeaturesMain {
 	}
 
 
-	private void checkNames(ObjectContainer container, Set names) {
-		Iterator iter = names.iterator();
-		while (iter.hasNext()) {
-			String name = (String)iter.next();
-			check(container, name);
-		}
+	private void checkNames(ObjectContainer container, Set expectedNames) {
+	    for (int i = 0; i < ALL_POSSIBLE_NAMES.length; i++) {
+            String name = ALL_POSSIBLE_NAMES[i];
+            check(container, name, expectedNames.contains(name));
+        }
 	}
 
-	private void checkClean() {
-		checkClean(_a);
-		checkClean(_b);
+	private void check(ObjectContainer container, String name, boolean isExpected) {
+        System.out.println(name + " " + isExpected);
+        Replicated obj = find(container, name);
+        if (isExpected) {
+            Test.ensure(obj != null);
+            container.delete(obj);
+        } else {
+            Test.ensure(obj == null);
+        }
 	}
 
-	private void checkClean(ObjectContainer container) {
-		Test.ensure(container.get(null).size() == 0);		
-	}
-
-	private void check(ObjectContainer container, String name) {
-		Query q = container.query();
-		q.constrain(Replicated.class);
-		q.descend("_name").constrain(name);
-		Object o = q.execute().next();
-		container.delete(o);
-		Test.ensure(!q.execute().hasNext());
-	}
-
+    private Replicated find(ObjectContainer container, String name) {
+        Query q = container.query();
+        q.constrain(Replicated.class);
+        q.descend("_name").constrain(name);
+        ObjectSet set = q.execute();
+        Test.ensure(set.size() < 2);
+        return (Replicated)set.next();
+    }
+    
 	private void initState() {
 		_a.set(new Replicated("a1"));
 		_a.set(new Replicated("a2"));
